+++ udata_front/theme/gouvfr/assets/js/components/AdminTable/AdminCommunityResourcesTable/AdminCommunityResourcesTable.vue
@@ -32,11 +32,25 @@
     <tbody>
       <tr v-for="communityResource in communityResources">
         <td>
-          <AdminContentWithTooltip>
+          <AdminContentWithTooltip class="fr-text--bold">
             <a class="fr-link fr-reset-link" :href="getCommunityResourceLinkToAdmin(communityResource)">
               <TextClamp :text="communityResource.title" :auto-resize="true" :max-lines="2"/>
             </a>
           </AdminContentWithTooltip>
+          <p v-if="datasets[communityResource.dataset.id]">
+            <a
+              class="fr-link inline-flex"
+              :href="datasets[communityResource.dataset.id]?.page"
+            >
+              <Vicon :height="12" class="self-center" :name="getSubjectTypeIcon('Dataset')"/>
+              <TextClamp
+                class="overflow-wrap-anywhere"
+                :text="datasets[communityResource.dataset.id]?.title"
+                :auto-resize="true"
+                :max-lines="1"
+              />
+            </a>
+          </p>
         </td>
         <td>
           <AdminBadge :type="getStatus(communityResource).type">{{ getStatus(communityResource).label }}</AdminBadge>
@@ -52,7 +66,8 @@
 </template>
 <script setup lang="ts">
 import { formatDate } from '@datagouv/components/ts';
-import type { CommunityResource } from '@datagouv/components/ts';
+import type { CommunityResource, Dataset } from '@datagouv/components/ts';
+import { OhVueIcon as Vicon } from "oh-vue-icons";
 import TextClamp from 'vue3-text-clamp';
 import { useI18n } from 'vue-i18n';
 import AdminBadge from '../../../components/AdminBadge/AdminBadge.vue';
@@ -61,6 +76,9 @@ import AdminTableTh from '../../../components/AdminTable/Table/AdminTableTh.vue'
 import AdminContentWithTooltip from '../../../components/AdminContentWithTooltip/AdminContentWithTooltip.vue';
 import { admin_root } from '../../../config';
 import type { AdminBadgeState, CommunityResourceSortedBy, SortDirection } from "../../../types";
+import { getSubjectTypeIcon } from '../../../api/discussions';
+import { ref, watchEffect } from 'vue';
+import { getDataset } from '../../../api/datasets';
 
 const props = defineProps<{
   communityResources: Array<CommunityResource>;
@@ -74,6 +92,20 @@ defineEmits<{
 }>();
 
 const { t } = useI18n();
+
+const datasetPromises = ref<Record<string, Promise<Dataset>>>({});
+const datasets = ref<Record<string, Dataset>>({});
+
+watchEffect(async () => {
+  for (const communityResource of props.communityResources) {
+    if(!(communityResource.dataset.id in datasetPromises.value)) {
+      const datasetPromise = getDataset(communityResource.dataset.id);
+      datasetPromises.value[communityResource.dataset.id] = datasetPromise;
+      const dataset = await datasetPromises.value[communityResource.dataset.id];
+      datasets.value[communityResource.dataset.id] = dataset;
+    }
+  }
+});
 
 function sorted(column: CommunityResourceSortedBy) {
   if(props.sortedBy === column) {